pyproject.toml
[project]
name = "prbuddy"
version = "0.1.0"
description = "AI-powered PR review companion that enables authors to train agents for reviewer Q&A"
requires-python = ">=3.10"

dependencies = [
    # OpenAI Agents SDK (includes realtime)
    "openai-agents>=0.6.4",
    "openai>=1.58.0",
    
    # MCP integration
    "mcp>=1.3.0",
    "httpx>=0.28.0",
    
    # Weaviate vector database
    "weaviate-client>=4.9.0",
    
    # AWS (Polly TTS)
    "boto3>=1.35.0",
    
    # ElevenLabs TTS (optional)
    "elevenlabs>=1.0.0",
    
    # Web server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    
    # Environment and utilities
    "python-dotenv>=1.0.1",
    
    # Configuration
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "ruamel.yaml>=0.18.0",
    
    # Observability
    "structlog>=24.0.0",
]

[project.optional-dependencies]
# Audio handling - only needed for local voice mode with mic/speaker
voice = [
    "pyaudio>=0.2.14",
    "numpy>=2.0.0",
]

# Development and testing dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests as integration tests (requires external services)",
]

